add test language switcher, component to test functionality
