+++ quiz/web/src/App.tsx
@@ -8,6 +8,7 @@ import FakerTest from "modules/genericTests/FakerTest/FakerTest";
 import FormikTest from "modules/genericTests/FormikTest/FormikTest";
 import TailwindTest from "modules/genericTests/TailwindTest/TailwindTest";
 import TranslationTest from "modules/genericTests/TranslationTest/TranslationTest";
+import LanguageSwitcher from "modules/genericTests/LanguageSwitcher/LanguageSwitcher";
 
 function App() {
     const queryClient = new QueryClient();
@@ -22,6 +23,7 @@ function App() {
                     <TailwindTest />
                     <FormikTest />
                     <TranslationTest />
+                    <LanguageSwitcher />
                 </ErrorBoundary>
             </AuthTokenProvider>
         </QueryClientProvider>
