+++ quiz/web/src/modules/common/i18nTranslation/languages/en.json
@@ -1,6 +1,9 @@
 {
   "translation": {
     "Learn": "Learn",
-    "Hello World": "Hello World"
+    "Hello World": "Hello World",
+    "Name": "Name",
+    "Message": "Message",
+    "Send": "Send"
   }
 }
