+++ quiz/web/src/modules/common/i18nTranslation/languages/pl.json
@@ -1,6 +1,9 @@
 {
   "translation": {
     "Learn": "Ucz się",
-    "Hello World": "Siemanko świecie"
+    "Hello World": "Siemanko świecie",
+    "Name": "Imię",
+    "Message": "Wiadomość",
+    "Send": "Wyślij"
   }
 }
