+++ quiz/web/src/modules/genericTests/TailwindTest/TailwindTest.tsx
@@ -1,5 +1,9 @@
+import { useTranslation } from "react-i18next";
+
 function TailwindTest() {
-    return <h1 className="text-3xl font-bold underline">Hello world!</h1>;
+    const { t } = useTranslation();
+
+    return <h1 className="text-3xl font-bold underline">{t("Hello World")}!</h1>;
 }
 
 export default TailwindTest;
